requirements.txt
google-cloud-bigquery
db-dtypes
Flask==2.2.3
gunicorn==20.1.0
python-telegram-bot==13.15
